test: put manifest-related test to separate test file
In former version the manifest handling was placed in config_file.c.
As this functionallity is now in a separate source file its tests should
be, too.
